+++ module-order/src/main/java/order/domain/OrderLineItem.java
@@ -1,0 +1,56 @@
+package order.domain;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+
+@Entity
+public class OrderLineItem {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long seq;
+
+    @ManyToOne
+    @JoinColumn(name = "order_id")
+    private Order order;
+
+    private Long menuId;
+    private long quantity;
+
+    public OrderLineItem() {
+    }
+
+    public OrderLineItem(Long menuId, long quantity) {
+        this(null, menuId, quantity);
+    }
+
+    public OrderLineItem(Long seq, Long menuId, long quantity) {
+        this.seq = seq;
+        this.menuId = menuId;
+        this.quantity = quantity;
+    }
+
+    public void setOrder(Order order) {
+        this.order = order;
+    }
+
+    public Long getSeq() {
+        return seq;
+    }
+
+    public Order getOrder() {
+        return order;
+    }
+
+    public Long getMenuId() {
+        return menuId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+}
