+++ module-order/src/main/java/order/domain/OrderLineItems.java
@@ -1,0 +1,25 @@
+package order.domain;
+
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.OneToMany;
+
+@Embeddable
+public class OrderLineItems {
+
+    @OneToMany(mappedBy = "order", cascade = {CascadeType.PERSIST, CascadeType.REMOVE})
+    private List<OrderLineItem> orderLineItems = new ArrayList<>();
+
+    public OrderLineItems() {
+    }
+
+    public OrderLineItems(List<OrderLineItem> orderLineItems) {
+        this.orderLineItems.addAll(orderLineItems);
+    }
+
+    public List<OrderLineItem> getOrderLineItems() {
+        return orderLineItems;
+    }
+}
