+++ module-order/src/main/java/order/dto/OrderLineItemRequestDto.java
@@ -1,0 +1,22 @@
+package order.dto;
+
+public class OrderLineItemRequestDto {
+    private Long menuId;
+    private long quantity;
+
+    public OrderLineItemRequestDto() {
+    }
+
+    public OrderLineItemRequestDto(Long menuId, long quantity) {
+        this.menuId = menuId;
+        this.quantity = quantity;
+    }
+
+    public Long getMenuId() {
+        return menuId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+}
