+++ module-order/src/main/java/order/repository/OrderRepository.java
@@ -1,0 +1,13 @@
+package order.repository;
+
+import common.domain.OrderStatus;
+import java.util.List;
+import order.domain.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+
+    boolean existsByOrderTableIdAndOrderStatusIn(Long orderTableId, List<OrderStatus> statusList);
+
+    boolean existsByOrderTableIdInAndOrderStatusIn(List<Long> orderTableIds, List<OrderStatus> statusList);
+}
