--- src/main/java/kitchenpos/order/application/TableValidatorImpl.java
@@ -1,38 +1,0 @@
-package kitchenpos.order.application;
-
-import java.util.Arrays;
-import java.util.Objects;
-import kitchenpos.common.domain.OrderStatus;
-import kitchenpos.common.exception.InvalidOrderStatusException;
-import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.table.application.TableValidator;
-import kitchenpos.table.domain.OrderTable;
-import org.springframework.stereotype.Service;
-
-@Service
-public class TableValidatorImpl implements TableValidator {
-
-    private final OrderRepository orderRepository;
-
-    public TableValidatorImpl(OrderRepository orderRepository) {
-        this.orderRepository = orderRepository;
-    }
-
-    public void checkValidChangeEmpty(OrderTable orderTable) {
-        checkNotGroup(orderTable);
-        checkOrderStatus(orderTable.getId());
-    }
-
-    private void checkNotGroup(OrderTable orderTable) {
-        if (Objects.nonNull(orderTable.getTableGroupId())) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void checkOrderStatus(Long orderTableId) {
-        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(orderTableId, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
-            throw new InvalidOrderStatusException();
-        }
-    }
-
-}
