--- src/main/java/kitchenpos/order/dto/OrderLineItemRequestDto.java
@@ -1,22 +1,0 @@
-package kitchenpos.order.dto;
-
-public class OrderLineItemRequestDto {
-    private Long menuId;
-    private long quantity;
-
-    public OrderLineItemRequestDto() {
-    }
-
-    public OrderLineItemRequestDto(Long menuId, long quantity) {
-        this.menuId = menuId;
-        this.quantity = quantity;
-    }
-
-    public Long getMenuId() {
-        return menuId;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-}
